+++ equalsverifier-release-verify/src/test/java/nl/jqno/equalsverifier/verify_release/jar/MainClassesJarTest.java
@@ -51,6 +51,16 @@ class MainClassesJarTest {
     }
 
     @Test
+    void assertAbsenceOfEmbeddedDependenciesUnderTheirOwnPath() {
+        jar.assertAbsenceOfEmbeddedDepedenciesUnderTheirOriginalPath();
+    }
+
+    @Test
+    void absenceOfOptionalDependencies() {
+        jar.assertAbsenceOfOptionalDependencies();
+    }
+
+    @Test
     void contentOfManifest() {
         jar.assertContentOfManifest("EqualsVerifier");
     }