+++ equalsverifier-release-verify/src/test/java/nl/jqno/equalsverifier/verify_release/jar/NodepClassesJarTest.java
@@ -49,6 +49,16 @@ class NodepClassesJarTest {
     }
 
     @Test
+    void assertAbsenceOfEmbeddedDependenciesUnderTheirOwnPath() {
+        jar.assertAbsenceOfEmbeddedDepedenciesUnderTheirOriginalPath();
+    }
+
+    @Test
+    void absenceOfOptionalDependencies() {
+        jar.assertAbsenceOfOptionalDependencies();
+    }
+
+    @Test
     void contentOfManifest() {
         jar.assertContentOfManifest("EqualsVerifier (no dependencies)");
     }