+++ equalsverifier-testhelpers/src/main/java/nl/jqno/equalsverifier_testhelpers/types/RecursiveTypeHelper.java
@@ -1,37 +1,210 @@
 package nl.jqno.equalsverifier_testhelpers.types;
 
+import java.util.Arrays;
+import java.util.List;
 import java.util.Objects;
 
 public class RecursiveTypeHelper {
 
-    public static final class Node {
+    public static class Node {
 
-        Node node;
+        final Node node;
+
+        public Node(Node node) {
+            this.node = node;
+        }
+
+        @Override
+        public final boolean equals(Object obj) {
+            return obj instanceof Node other && Objects.equals(node, other.node);
+        }
+
+        @Override
+        public final int hashCode() {
+            return Objects.hash(node);
+        }
+    }
+
+    public static class SubNode extends Node {
+
+        public SubNode(Node node) {
+            super(node);
+        }
+    }
+
+    public static class NodeContainer {
+
+        final Node node;
+
+        public NodeContainer(Node node) {
+            this.node = node;
+        }
+
+        @Override
+        public final boolean equals(Object obj) {
+            return obj instanceof NodeContainer other && Objects.equals(node, other.node);
+        }
+
+        @Override
+        public final int hashCode() {
+            return Objects.hash(node);
+        }
+    }
+
+    public static class SubNodeContainer extends NodeContainer {
+
+        public SubNodeContainer(Node node) {
+            super(node);
+        }
+    }
+
+    public static class Tree {
+
+        final List<Tree> innerTrees;
+
+        public Tree(List<Tree> innerTrees) {
+            this.innerTrees = innerTrees;
+        }
+
+        @Override
+        public final boolean equals(Object obj) {
+            return obj instanceof Tree other && Objects.equals(innerTrees, other.innerTrees);
+        }
+
+        @Override
+        public final int hashCode() {
+            return Objects.hash(innerTrees);
+        }
+    }
+
+    public static class TreeContainer {
+
+        final Tree tree;
+
+        public TreeContainer(Tree tree) {
+            this.tree = tree;
+        }
+
+        @Override
+        public final boolean equals(Object obj) {
+            return obj instanceof TreeContainer other && Objects.equals(tree, other.tree);
+        }
+
+        @Override
+        public final int hashCode() {
+            return Objects.hash(tree);
+        }
+    }
+
+    @SuppressWarnings("unused")
+    public static final class StaticFinalNodeContainer {
+
+        private static final Node NODE = new Node(null);
+        private final int i;
+
+        public StaticFinalNodeContainer(int i) {
+            this.i = i;
+        }
+
+        @Override
+        public boolean equals(Object obj) {
+            return obj instanceof StaticFinalNodeContainer other && Objects.equals(i, other.i);
+        }
+
+        @Override
+        public int hashCode() {
+            return Objects.hash(i);
+        }
     }
 
     public static final class NodeArray {
 
-        NodeArray[] nodeArrays;
+        final NodeArray[] nodeArray;
+
+        public NodeArray(NodeArray[] nodeArray) {
+            this.nodeArray = nodeArray;
+        }
+
+        @Override
+        public boolean equals(Object obj) {
+            return obj instanceof NodeArray other && Arrays.equals(nodeArray, other.nodeArray);
+        }
+
+        @Override
+        public int hashCode() {
+            return Arrays.hashCode(nodeArray);
+        }
     }
 
     public static final class TwoStepNodeA {
 
-        TwoStepNodeB node;
+        final TwoStepNodeB node;
+
+        public TwoStepNodeA(TwoStepNodeB node) {
+            this.node = node;
+        }
+
+        @Override
+        public boolean equals(Object obj) {
+            return obj instanceof TwoStepNodeA other && Objects.equals(node, other.node);
+        }
+
+        @Override
+        public int hashCode() {
+            return Objects.hash(node);
+        }
     }
 
     public static final class TwoStepNodeB {
 
-        TwoStepNodeA node;
+        final TwoStepNodeA node;
+
+        public TwoStepNodeB(TwoStepNodeA node) {
+            this.node = node;
+        }
+
+        @Override
+        public boolean equals(Object obj) {
+            return obj instanceof TwoStepNodeB other && Objects.equals(node, other.node);
+        }
     }
 
     public static final class TwoStepNodeArrayA {
 
-        TwoStepNodeArrayB[] nodes;
+        final TwoStepNodeArrayB[] nodes;
+
+        public TwoStepNodeArrayA(TwoStepNodeArrayB[] nodes) {
+            this.nodes = nodes;
+        }
+
+        @Override
+        public boolean equals(Object obj) {
+            return obj instanceof TwoStepNodeArrayA other && Arrays.equals(nodes, other.nodes);
+        }
+
+        @Override
+        public int hashCode() {
+            return Arrays.hashCode(nodes);
+        }
     }
 
     public static final class TwoStepNodeArrayB {
 
-        TwoStepNodeArrayA[] nodes;
+        final TwoStepNodeArrayA[] nodes;
+
+        public TwoStepNodeArrayB(TwoStepNodeArrayA[] nodes) {
+            this.nodes = nodes;
+        }
+
+        @Override
+        public boolean equals(Object obj) {
+            return obj instanceof TwoStepNodeArrayB other && Arrays.equals(nodes, other.nodes);
+        }
+
+        @Override
+        public int hashCode() {
+            return Arrays.hashCode(nodes);
+        }
     }
 
     public static final class NotRecursiveA {
@@ -70,11 +243,7 @@ public class RecursiveTypeHelper {
 
         @Override
         public boolean equals(Object obj) {
-            if (!(obj instanceof RecursiveType)) {
-                return false;
-            }
-            var other = (RecursiveType) obj;
-            return Objects.equals(recurse, other.recurse);
+            return obj instanceof RecursiveType other && Objects.equals(recurse, other.recurse);
         }
 
         @Override
@@ -93,11 +262,7 @@ public class RecursiveTypeHelper {
 
         @Override
         public boolean equals(Object obj) {
-            if (!(obj instanceof RecursiveTypeContainer)) {
-                return false;
-            }
-            var other = (RecursiveTypeContainer) obj;
-            return Objects.equals(recurse, other.recurse);
+            return obj instanceof RecursiveTypeContainer other && Objects.equals(recurse, other.recurse);
         }
 
         @Override